requirements.txt
streamlit==1.35.0
numpy==1.26
scikit-learn==1.5.0
scipy==1.13.1
regex==2024.5.15
requests==2.32.3
transformers==4.42.3
torch==2.5.1+cu121
beautifulsoup4==4.11.2
chromadb==0.6.3
PyMuPDF==1.25.3
pandas==2.1.4
pywin32==306
python-dotenv==1.0.1